Add G13 functionality :D
If a submission is eligible, we show a “Request G13 deletion” button in
the main panel. When clicked, the script will add {{db-g13}} to the
submission and notify the page creator as well as all users who have
submitted the page for review in the past. After notifying, we log the
G13’ing for good measure to the user’s CSD log.

- AFCH.dateToMwTimestamp() function added to core