"Recently used" decline reasons! (yay, mw.user.options <3)
Add AFCH.userData as a wrapper around mw.user.options. The main benefit
of using this wrapper is when the page is reloaded using ajax — we use
an intermediate cache which stores values until mw.user.options updates
(on reload). Additionally, handles the API requests to set values
(which is nice ;) ).

As far as the “recently used” menu goes, it shows the three most
recently selected decline reasons as well as the “custom” reason (which
is always displayed).