AFCH.userData: add localStorage layer
The reason for this redundancy is because of an obnoxious
little thing called caching. Ideally the script would simply
use mw.user.options, but *apparently* MediaWiki doesn't always provide
the most updated mw.user.options on page load -- in some instances, it
will provide an stale, cached version instead. This is most certainly a
MediaWiki bug, but in the meantime, we circumvent it by adding numerous
layers of redundancy to the whole getup. In this manner, hopefully by
the time we have to rely on mw.user.options, the cache will have been
invalidated and the world won't explode. *sighs repeatedly*