Add prefillBiographyDetails which may save some time; hygiene
Currently prefillBiographyDetails prefills the LISTAS/DEFAULTSORT if
the page title is only two words—assumes that word 1 is first name,
word 2 is last name. This won’t be right 100% of the title, but for the
most part might be kinda handy.

In the future, envision also having the pre-existing PERSONDATA
detection here…