Update Jest to use 0.5.x branch; fix #65
Jest doesn’t work with Node 0.12.x (due to jsdom woes), but the 0.5.x
branch on GitHub has been updated and works with the latest version of
iojs. So, from now on, iojs must be used to build afch-rewrite (a la
nvm, probably). This should be added to the README/wiki…