Remove minification and vendor-prefixing while building LESS files (#431)
* Remove grunt-autoprefixer

The only CSS rule requiring vendor-prefixing is user-select: none. It's supported everywhere except for Safari which needs the -webkit- prefix. Added it manually and removed the grunt-autoprefixer dependency. All other CSS being used in the gadget is standard.

* Remove css minification

Remove grunt-cssmin dependency. CSS minification is unnecessary in the build as it anyway gets handled by ResourceLoader. This also makes the on-wiki CSS file readable.

* Remove minimal.less: empty and unused