submissions: shorten summary for custom declines
When you have a very long custom decline reason, this would be included
in your edit summary, up to a certain limit. This can look a bit silly
as it sort of defeats the point of the edit summary being a summary, so
if we were to truncate the edit summary due to a long custom decline
rationale, we then just don't mention it at all.